replace old components with new ones
