+++ src/app/Instance/CreateInstance/components/ErrorHandlingSelection.tsx
@@ -6,7 +6,7 @@ import {
   SelectVariant,
 } from "@patternfly/react-core";
 import { useTranslation } from "@rhoas/app-services-ui-components";
-import { ErrorHandlingMethods } from "../../../types/ErrorHandlingMethods";
+import { ErrorHandlingMethods } from "../../../../types/ErrorHandlingMethods";
 
 interface ErrorHandlingSelectionProps {
   defaultMethod: string;
