+++ src/types/Error.ts
@@ -1,11 +1,3 @@
-export interface ResponseError {
-  kind: "Error";
-  id: string;
-  href: string;
-  code: string;
-  reason: string;
-}
-
 export class ErrorWithDetail extends Error {
   detailSection?: JSX.Element;
   constructor(detailSection?: JSX.Element, message?: string) {
